requirements.txt
flask
flask-cors
groq
edge-tts
python-dotenv
gunicorn
nest_asyncio
